requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environ==0.10.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rosetta==0.9.8
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jdatetime==4.1.0
oauthlib==3.2.2
Pillow==9.3.0
polib==1.1.1
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
